requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# NLP Libraries
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Finnish NLP specific
voikko==0.5.2
ufal.udpipe==1.3.0.1

# Text Processing
nltk==3.8.1
regex==2023.12.25

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Frontend
streamlit==1.30.0

# Database (Optional)
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# CORS
python-cors==1.0.0
